// src/app/models/Persona.ts
export interface Persona{
    id: number;
    nombre: string;
    apellidopaterno: string;
    apellidomaterno: string;
    edad: number;
    estadocivil: string;
    telefono: string;
    comosupistedeiglesia: string;
    colonia: string;
    nombrevoluntario: string;
    condicionvisita: string;
    grupopequeñointeres: string;
    oraciondefe: boolean;
}